Struct iced_glow::Command

pub struct Command<T>(_);

A set of asynchronous actions to be performed by some runtime.

Implementations

impl<T> Command<T>

pub const fn none() -> Command<T>

Creates an empty Command.

In other words, a Command that does nothing.

pub const fn single(action: Action<T>) -> Command<T>

Creates a Command that performs a single [Action].

pub fn widget(operation: impl Operation<T> + 'static) -> Command<T>

Creates a Command that performs a [widget::Operation].

pub fn perform<A>(
future: impl Future<Output = T> + MaybeSend + 'static,
f: impl FnOnce(T) -> A + MaybeSend + 'static
) -> Command<A>

Creates a Command that performs the action of the given future.

pub fn batch(commands: impl IntoIterator<Item = Command<T>>) -> Command<T>

Creates a Command that performs the actions of all the given commands.

Once this command is run, all the commands will be executed at once.

pub fn map<A>(
self,
f: impl Fn(T) -> A + MaybeSend + Sync + Clone + 'static
) -> Command<A>where
T: 'static,
A: 'static,

Applies a transformation to the result of a Command.

pub fn actions(self) -> Vec<Action<T>, Global>

Returns all of the actions of the Command.