Trait iced_glow::widget::canvas::Program

pub trait Program<Message> {
    fn draw(
        &self, 
        bounds: Rectangle<f32>, 
        cursor: Cursor
    ) -> Vec<Geometry, Global>;

    fn update(
        &mut self, 
        _event: Event, 
        _bounds: Rectangle<f32>, 
        _cursor: Cursor
    ) -> (Status, Option<Message>) { ... }

    fn mouse_interaction(
        &self, 
        _bounds: Rectangle<f32>, 
        _cursor: Cursor
    ) -> Interaction { ... }
}

The state and logic of a Canvas.

A Program can mutate internal state and produce messages for an application.

Required methods

fn draw(&self, bounds: Rectangle<f32>, cursor: Cursor) -> Vec<Geometry, Global>

Draws the state of the Program, producing a bunch of Geometry.

Geometry can be easily generated with a Frame or stored in a Cache.

Provided methods

fn update(
&mut self,
_event: Event,
_bounds: Rectangle<f32>,
_cursor: Cursor
) -> (Status, Option<Message>)

Updates the state of the Program.

When a Program is used in a Canvas, the runtime will call this method for each Event.

This method can optionally return a Message to notify an application of any meaningful interactions.

By default, this method does and returns nothing.

fn mouse_interaction(
&self,
_bounds: Rectangle<f32>,
_cursor: Cursor
) -> Interaction

Returns the current mouse interaction of the Program.

The interaction returned will be in effect even if the cursor position is out of bounds of the program’s Canvas.